Reorganization of content in action support headers
